mi5: move patches to a separate module
